Set relay1 high on speed medium
On the original implementation from Sonoff, the medium speed have both relays 1 and 2 on HIGH, which makes sense, as those two have the same capacitance.
With the previous implementation (only relay 2 HIGH) there's no difference between speed low and speed medium, as the capacitors and therefore the resulting voltage are the same when only relay 1 or only relay 2.